move DateInterval to utils
